Added "enabled" config setting and fixed a plugin dependency in BuildConfig.
Previously, there was no way to have this plugin installed in an
application, but unused. With the plugin installed, but Redis not
available, a series of exceptions, including JedisConnectionException,
were thrown when starting the application. I added the "enabled" config
setting which makes it possible for an application to have this plugin
installed, but choose not to use it.

I also modified BuildConfig to make the Hibernate plugin a "compile"
dependency rather than a "test" dependency. Without this change the
plugin would not compile for me.